separate classes into separate files
each file containing a type of class
